Conntrack map upgrade changes
